refactor: move Point into PointShapes
