+++ src/RogueLike/Main/AoE/Line.java
@@ -1,5 +1,7 @@
 package RogueLike.Main.AoE;
 
+import RogueLike.Main.Utils.PointShapes.Point;
+
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
