+++ src/RogueLike/Main/Entities/Entity.java
@@ -1,6 +1,6 @@
 package RogueLike.Main.Entities;
 
-import RogueLike.Main.AoE.Point;
+import RogueLike.Main.Utils.PointShapes.Point;
 import RogueLike.Main.Creatures.Creature;
 import RogueLike.Main.World;
