+++ src/RogueLike/Main/FieldOfView.java
@@ -1,7 +1,7 @@
 package RogueLike.Main;
 
 import RogueLike.Main.AoE.Line;
-import RogueLike.Main.AoE.Point;
+import RogueLike.Main.Utils.PointShapes.Point;
 
 public class FieldOfView {
     private World world;
