+++ src/RogueLike/Main/Screens/ThrowAtScreen.java
@@ -1,7 +1,7 @@
 package RogueLike.Main.Screens;
 
 import RogueLike.Main.AoE.Line;
-import RogueLike.Main.AoE.Point;
+import RogueLike.Main.Utils.PointShapes.Point;
 import RogueLike.Main.Creatures.Creature;
 import RogueLike.Main.Items.Item;
