+++ src/RogueLike/Main/Worldgen/Blueprints/MerchantFloor.java
@@ -1,7 +1,7 @@
 package RogueLike.Main.Worldgen.Blueprints;
 
 import RogueLike.Main.AoE.Line;
-import RogueLike.Main.AoE.Point;
+import RogueLike.Main.Utils.PointShapes.Point;
 import RogueLike.Main.Creatures.Creature;
 import RogueLike.Main.Entities.Teleporter;
 import RogueLike.Main.Tile;
