+++ src/RogueLike/Main/Worldgen/Structures/PitItem.java
@@ -1,6 +1,6 @@
 package RogueLike.Main.Worldgen.Structures;
 
-import RogueLike.Main.AoE.Point;
+import RogueLike.Main.Utils.PointShapes.Point;
 import RogueLike.Main.Items.Item;
 import RogueLike.Main.Tile;
 import RogueLike.Main.Utils.PointShapes.Rectangle;
