+++ src/RogueLike/Main/Worldgen/Structures/SmallCell.java
@@ -1,8 +1,6 @@
 package RogueLike.Main.Worldgen.Structures;
 
-import RogueLike.Main.AoE.Point;
-import RogueLike.Main.Tile;
-import RogueLike.Main.Utils.PointShapes.HollowRectangle;
+import RogueLike.Main.Utils.PointShapes.Point;
 import RogueLike.Main.Utils.PointShapes.Rectangle;
 import RogueLike.Main.World;
 import RogueLike.Main.WorldBuilder;
